Refactor CreateFilter method to generic method and move to separate file. Update queryhandlers.
