+++ src/Application/Books/Author/GetAll/GetAuthorsQueryHandler.cs
@@ -1,12 +1,11 @@
 ﻿using Application.Abstractions.Data;
 using Application.Abstractions.Messaging;
-using Application.Dto;
 using Application.Dto.Books;
 using Application.Dto.ConvertObjects;
+using Application.Extensions;
 using Domain.Models.Books;
 using Microsoft.EntityFrameworkCore;
 using SharedKernel;
-using System.Linq.Expressions;
 
 namespace Application.Books
 {
@@ -40,7 +39,7 @@ namespace Application.Books
             }
             else
             {
-                authorsQuery = context.AuthorRepository.GetAll(CreateFilter(request.SearchParams));
+                authorsQuery = context.AuthorRepository.GetAll(ExpressionBuilder.CreateFilter<Author>(request.SearchParams));
             }
 
             IReadOnlyList<Author>? authors = await authorsQuery.Skip(skip).Take(request.SearchParams.PageSize).ToListAsync(cancellationToken);
@@ -55,39 +54,6 @@ namespace Application.Books
 
             // Return a success result with the list of author DTOs.
             return Result.Success(authorDtos);
-
         }
-
-        private static Expression<Func<Author, bool>> CreateFilter(SearchParamsDTO searchParams)
-        {
-            var parameter = Expression.Parameter(typeof(Author), "genre");
-            Expression body = Expression.Equal(
-                Expression.Property(parameter, nameof(Author.IsActive)),
-                Expression.Constant(true)
-            );
-
-            foreach (var filter in searchParams.SearchParams)
-            {
-                var propertyInfo = typeof(Author).GetProperty(filter.PropertyName);
-                if (propertyInfo == null || propertyInfo.PropertyType != typeof(string))
-                    continue;
-
-                var property = Expression.Property(parameter, filter.PropertyName);
-                var value = Expression.Constant(filter.Value, typeof(string));
-
-                Expression filterExpr = filter.MatchType switch
-                {
-                    SearchType.Contains => Expression.Call(property, nameof(string.Contains), null, value),
-                    SearchType.Exact => Expression.Equal(property, value),
-                    SearchType.StartsWith => Expression.Call(property, nameof(string.StartsWith), null, value),
-                    SearchType.EndsWith => Expression.Call(property, nameof(string.EndsWith), null, value),
-                    _ => Expression.Constant(true)
-                };
-                body = Expression.AndAlso(body, filterExpr);
-            }
-
-            return Expression.Lambda<Func<Author, bool>>(body, parameter);
-        }
-
     }
 }
