+++ src/Application/Common/Genre/GetAll/GetGenresQueryHandler.cs
@@ -4,10 +4,8 @@ using SharedKernel;
 using Application.Abstractions.Data;
 using Application.Dto.ConvertObjects;
 using Domain.Models.Common;
-using System.Linq.Expressions;
-using Application.Dto;
-using SharedKernel.Extensions;
 using Microsoft.EntityFrameworkCore;
+using Application.Extensions;
 
 namespace Application.Common
 {
@@ -37,7 +35,7 @@ namespace Application.Common
             }
             else
             {
-                genresQuery = context.GenreRepository.GetAll(CreateFilter(request.SearchParams));
+                genresQuery = context.GenreRepository.GetAll(ExpressionBuilder.CreateFilter<Genre>(request.SearchParams));
             }
 
             IReadOnlyList<Genre>? genres = await genresQuery.Skip(skip).Take(request.SearchParams.PageSize).ToListAsync(cancellationToken);
@@ -52,37 +50,6 @@ namespace Application.Common
 
             // Return a success result with the list of genre DTOs.
             return Result.Success(genreDtos);
-        }
-
-        private static Expression<Func<Genre, bool>> CreateFilter(SearchParamsDTO searchParams)
-        {
-            var parameter = Expression.Parameter(typeof(Genre), "genre");
-            Expression body = Expression.Equal(
-                Expression.Property(parameter, nameof(Genre.IsActive)),
-                Expression.Constant(true)
-            );
-
-            foreach(var filter in searchParams.SearchParams)
-            {
-                var propertyInfo = typeof(Genre).GetProperty(filter.PropertyName);
-                if (propertyInfo == null || propertyInfo.PropertyType != typeof(string))
-                    continue;
-
-                var property = Expression.Property(parameter, filter.PropertyName);
-                var value = Expression.Constant(filter.Value, typeof(string));
-
-                Expression filterExpr = filter.MatchType switch
-                {
-                    SearchType.Contains => Expression.Call(property, nameof(string.Contains), null, value),
-                    SearchType.Exact => Expression.Equal(property, value),
-                    SearchType.StartsWith => Expression.Call(property, nameof(string.StartsWith), null, value),
-                    SearchType.EndsWith => Expression.Call(property, nameof(string.EndsWith), null, value),
-                    _ => Expression.Constant(true)
-                };
-                body = Expression.AndAlso(body, filterExpr);
-            }
-
-            return Expression.Lambda<Func<Genre, bool>>(body,parameter);
         }
     }
 }
