+++ src/Application/Common/Language/GetAll/GetLanguagesQueryHandler.cs
@@ -1,13 +1,11 @@
 using Application.Abstractions.Data;
 using Application.Abstractions.Messaging;
-using Application.Dto;
 using Application.Dto.Common;
 using Application.Dto.ConvertObjects;
+using Application.Extensions;
 using Domain.Models.Common;
 using Microsoft.EntityFrameworkCore;
 using SharedKernel;
-using SharedKernel.Extensions;
-using System.Linq.Expressions;
 
 namespace Application.Common
 {
@@ -37,7 +35,7 @@ namespace Application.Common
             }
             else
             {
-                languagesQuery = context.LanguageRepository.GetAll(CreateFilter(request.SearchParams));
+                languagesQuery = context.LanguageRepository.GetAll(ExpressionBuilder.CreateFilter<Language>(request.SearchParams));
             }
             // Retrieve all languages from the repository.
             IReadOnlyList<Language> languages = await languagesQuery.Skip(skip).Take(request.SearchParams.PageSize).ToListAsync(cancellationToken);
@@ -53,32 +51,6 @@ namespace Application.Common
 
             // Return the successful result with the list of language DTOs.
             return Result.Success(languageDtos);
-        }
-
-        private static Expression<Func<Language, bool>> CreateFilter(SearchParamsDTO searchParams)
-        {
-            Expression<Func<Language, bool>> predicate = genre => genre.IsActive;
-            foreach (var filter in searchParams.SearchParams)
-            {
-                Expression<Func<Language, bool>> filterExpr = filter.MatchType switch
-                {
-                    SearchType.Contains => language =>
-                        (language.GetPropertyValue(filter.PropertyName)!.ToString() ?? string.Empty)
-                            .Contains(filter.Value),
-                    SearchType.Exact => language =>
-                        (language.GetPropertyValue(filter.PropertyName)!.ToString() ?? string.Empty)
-                            == filter.Value,
-                    SearchType.StartsWith => language =>
-                        (language.GetPropertyValue(filter.PropertyName)!.ToString() ?? string.Empty)
-                            .StartsWith(filter.Value),
-                    SearchType.EndsWith => language =>
-                        (language.GetPropertyValue(filter.PropertyName)!.ToString() ?? string.Empty)
-                            .EndsWith(filter.Value),
-                    _ => language => true
-                };
-                predicate = predicate.AndAlso(filterExpr);
-            }
-            return predicate;
         }
     }
 }
